+++ query/match.c
@@ -2,23 +2,21 @@
  * Licensed under the Apache License, Version 2.0 */
 
 #include "watchman.h"
+#include "thirdparty/wildmatch/wildmatch.h"
 
-#ifndef FNM_CASEFOLD
-# define FNM_CASEFOLD 0
-# define NO_CASELESS_FNMATCH 1
-#endif
-
-struct match_data {
+struct wildmatch_data {
   char *pattern;
   bool caseless;
   bool wholename;
+  bool noescape;
+  bool includedotfiles;
 };
 
-static bool eval_match(struct w_query_ctx *ctx,
+static bool eval_wildmatch(struct w_query_ctx *ctx,
     struct watchman_file *file,
     void *data)
 {
-  struct match_data *match = data;
+  struct wildmatch_data *match = data;
   w_string_t *str;
 
   if (match->wholename) {
@@ -27,36 +25,44 @@ static bool eval_match(struct w_query_ctx *ctx,
     str = file->name;
   }
 
-  return fnmatch(match->pattern,
-      str->buf,
-      FNM_PERIOD |
-      (match->caseless ? FNM_CASEFOLD : 0)) == 0;
+  return wildmatch(
+    match->pattern,
+    str->buf,
+    (match->includedotfiles ? 0 : WM_PERIOD) |
+    (match->noescape ? WM_NOESCAPE : 0) |
+    (match->wholename ? WM_PATHNAME : 0) |
+    (match->caseless ? WM_CASEFOLD : 0),
+    0) == WM_MATCH;
 }
 
 static void dispose_match(void *data)
 {
-  struct match_data *match = data;
+  struct wildmatch_data *match = data;
 
   free(match->pattern);
   free(match);
 }
 
-static w_query_expr *match_parser_inner(w_query *query,
+static w_query_expr *wildmatch_parser_inner(w_query *query,
     json_t *term, bool caseless)
 {
   const char *ignore, *pattern, *scope = "basename";
   const char *which = caseless ? "imatch" : "match";
-  struct match_data *data;
+  int noescape = 0;
+  int includedotfiles = 0;
+  struct wildmatch_data *data;
 
-#ifdef NO_CASELESS_FNMATCH
-  if (caseless) {
-    asprintf(&query->errmsg,
-        "imatch: Your system doesn't support FNM_CASEFOLD");
-    return NULL;
-  }
-#endif
-
-  if (json_unpack(term, "[s,s,s]", &ignore, &pattern, &scope) != 0 &&
+  if (json_unpack(
+        term,
+        "[s,s,s,{s?b,s?b,s?b}]",
+        &ignore,
+        &pattern,
+        &scope,
+        "noescape",
+        &noescape,
+        "includedotfiles",
+        &includedotfiles) != 0 &&
+      json_unpack(term, "[s,s,s]", &ignore, &pattern, &scope) != 0 &&
       json_unpack(term, "[s,s]", &ignore, &pattern) != 0) {
     ignore_result(asprintf(&query->errmsg,
         "Expected [\"%s\", \"pattern\", \"scope\"?]",
@@ -75,21 +81,23 @@ static w_query_expr *match_parser_inner(w_query *query,
   data->pattern = strdup(pattern);
   data->caseless = caseless;
   data->wholename = !strcmp(scope, "wholename");
+  data->noescape = noescape;
+  data->includedotfiles = includedotfiles;
 
-  return w_query_expr_new(eval_match, dispose_match, data);
+  return w_query_expr_new(eval_wildmatch, dispose_match, data);
 }
 
-static w_query_expr *match_parser(w_query *query, json_t *term)
+static w_query_expr *wildmatch_parser(w_query *query, json_t *term)
 {
-  return match_parser_inner(query, term, !query->case_sensitive);
+  return wildmatch_parser_inner(query, term, !query->case_sensitive);
 }
-W_TERM_PARSER("match", match_parser)
+W_TERM_PARSER("match", wildmatch_parser)
 
-static w_query_expr *imatch_parser(w_query *query, json_t *term)
+static w_query_expr *iwildmatch_parser(w_query *query, json_t *term)
 {
-  return match_parser_inner(query, term, true);
+  return wildmatch_parser_inner(query, term, true);
 }
-W_TERM_PARSER("imatch", imatch_parser)
+W_TERM_PARSER("imatch", iwildmatch_parser)
 
 /* vim:ts=2:sw=2:et:
  */